Rename test projects for clarity.
