+++ Northwind.sln
@@ -19,11 +19,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Domain", "Src\Domain\Domain
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Application", "Src\Application\Application.csproj", "{5FB1A4D7-D13E-47C3-9152-3890B8CA9C37}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Application.Tests", "Tests\Application.Tests\Application.Tests.csproj", "{C379031A-C615-4F06-BBAB-B27196335FB4}"
-	ProjectSection(ProjectDependencies) = postProject
-		{5FB1A4D7-D13E-47C3-9152-3890B8CA9C37} = {5FB1A4D7-D13E-47C3-9152-3890B8CA9C37}
-	EndProjectSection
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Docs", "Docs", "{36DA64CD-41D5-48FD-8AFE-EBBF8EFD57ED}"
 	ProjectSection(SolutionItems) = preProject
 		README.md = README.md
@@ -32,19 +27,15 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Docs", "Docs", "{36DA64CD-4
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Infrastructure", "Src\Infrastructure\Infrastructure.csproj", "{1C11E54E-54DB-47A2-9D10-03AFDA347F22}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Domain.Tests", "Tests\Domain.Tests\Domain.Tests.csproj", "{73C1627E-BB1D-4159-A418-3F4F5BA7820C}"
-	ProjectSection(ProjectDependencies) = postProject
-		{0A18018C-95CA-43CB-BFBF-5DB086E93009} = {0A18018C-95CA-43CB-BFBF-5DB086E93009}
-	EndProjectSection
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Persistence", "Src\Persistence\Persistence.csproj", "{34BB0A51-74B9-46B6-BEAA-94A914F552EE}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WebUI", "Src\WebUI\WebUI.csproj", "{B3A105C1-F3E0-4290-A330-D0D3F29D2B26}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WebUI.FunctionalTests", "Tests\WebUI.FunctionalTests\WebUI.FunctionalTests.csproj", "{A5A16118-2C05-4433-8C39-384AB81E07CA}"
-	ProjectSection(ProjectDependencies) = postProject
-		{B3A105C1-F3E0-4290-A330-D0D3F29D2B26} = {B3A105C1-F3E0-4290-A330-D0D3F29D2B26}
-	EndProjectSection
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Application.UnitTests", "Tests\Application.UnitTests\Application.UnitTests.csproj", "{69E657A0-DE8D-4BA4-B9A9-77C8332321C6}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Domain.UnitTests", "Tests\Domain.UnitTests\Domain.UnitTests.csproj", "{FD68C950-322B-467C-82A6-99C084BAB50B}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WebUI.IntegrationTests", "Tests\WebUI.IntegrationTests\WebUI.IntegrationTests.csproj", "{22B85BF9-27F9-4C0C-AB6D-C05014C022E5}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -64,18 +55,10 @@ Global
 		{5FB1A4D7-D13E-47C3-9152-3890B8CA9C37}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5FB1A4D7-D13E-47C3-9152-3890B8CA9C37}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{5FB1A4D7-D13E-47C3-9152-3890B8CA9C37}.Release|Any CPU.Build.0 = Release|Any CPU
-		{C379031A-C615-4F06-BBAB-B27196335FB4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C379031A-C615-4F06-BBAB-B27196335FB4}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C379031A-C615-4F06-BBAB-B27196335FB4}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C379031A-C615-4F06-BBAB-B27196335FB4}.Release|Any CPU.Build.0 = Release|Any CPU
 		{1C11E54E-54DB-47A2-9D10-03AFDA347F22}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{1C11E54E-54DB-47A2-9D10-03AFDA347F22}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{1C11E54E-54DB-47A2-9D10-03AFDA347F22}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{1C11E54E-54DB-47A2-9D10-03AFDA347F22}.Release|Any CPU.Build.0 = Release|Any CPU
-		{73C1627E-BB1D-4159-A418-3F4F5BA7820C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{73C1627E-BB1D-4159-A418-3F4F5BA7820C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{73C1627E-BB1D-4159-A418-3F4F5BA7820C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{73C1627E-BB1D-4159-A418-3F4F5BA7820C}.Release|Any CPU.Build.0 = Release|Any CPU
 		{34BB0A51-74B9-46B6-BEAA-94A914F552EE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{34BB0A51-74B9-46B6-BEAA-94A914F552EE}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{34BB0A51-74B9-46B6-BEAA-94A914F552EE}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -84,10 +67,18 @@ Global
 		{B3A105C1-F3E0-4290-A330-D0D3F29D2B26}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B3A105C1-F3E0-4290-A330-D0D3F29D2B26}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B3A105C1-F3E0-4290-A330-D0D3F29D2B26}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A5A16118-2C05-4433-8C39-384AB81E07CA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A5A16118-2C05-4433-8C39-384AB81E07CA}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A5A16118-2C05-4433-8C39-384AB81E07CA}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A5A16118-2C05-4433-8C39-384AB81E07CA}.Release|Any CPU.Build.0 = Release|Any CPU
+		{69E657A0-DE8D-4BA4-B9A9-77C8332321C6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{69E657A0-DE8D-4BA4-B9A9-77C8332321C6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{69E657A0-DE8D-4BA4-B9A9-77C8332321C6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{69E657A0-DE8D-4BA4-B9A9-77C8332321C6}.Release|Any CPU.Build.0 = Release|Any CPU
+		{FD68C950-322B-467C-82A6-99C084BAB50B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{FD68C950-322B-467C-82A6-99C084BAB50B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{FD68C950-322B-467C-82A6-99C084BAB50B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{FD68C950-322B-467C-82A6-99C084BAB50B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{22B85BF9-27F9-4C0C-AB6D-C05014C022E5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{22B85BF9-27F9-4C0C-AB6D-C05014C022E5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{22B85BF9-27F9-4C0C-AB6D-C05014C022E5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{22B85BF9-27F9-4C0C-AB6D-C05014C022E5}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -99,12 +90,12 @@ Global
 		{3C7E8164-A2CE-471A-9C15-E1A74E4BE9D9} = {0FEB6ADF-29FD-48CE-AE0F-9E19E9209795}
 		{0A18018C-95CA-43CB-BFBF-5DB086E93009} = {0FEB6ADF-29FD-48CE-AE0F-9E19E9209795}
 		{5FB1A4D7-D13E-47C3-9152-3890B8CA9C37} = {0FEB6ADF-29FD-48CE-AE0F-9E19E9209795}
-		{C379031A-C615-4F06-BBAB-B27196335FB4} = {9FEDAEFB-297C-4201-8FAA-F502F94CAFED}
 		{1C11E54E-54DB-47A2-9D10-03AFDA347F22} = {1EC1EEFF-3888-4125-9229-623C564F3E7E}
-		{73C1627E-BB1D-4159-A418-3F4F5BA7820C} = {9FEDAEFB-297C-4201-8FAA-F502F94CAFED}
 		{34BB0A51-74B9-46B6-BEAA-94A914F552EE} = {1EC1EEFF-3888-4125-9229-623C564F3E7E}
 		{B3A105C1-F3E0-4290-A330-D0D3F29D2B26} = {833DEA57-A970-418C-831C-2355B8294766}
-		{A5A16118-2C05-4433-8C39-384AB81E07CA} = {9FEDAEFB-297C-4201-8FAA-F502F94CAFED}
+		{69E657A0-DE8D-4BA4-B9A9-77C8332321C6} = {9FEDAEFB-297C-4201-8FAA-F502F94CAFED}
+		{FD68C950-322B-467C-82A6-99C084BAB50B} = {9FEDAEFB-297C-4201-8FAA-F502F94CAFED}
+		{22B85BF9-27F9-4C0C-AB6D-C05014C022E5} = {9FEDAEFB-297C-4201-8FAA-F502F94CAFED}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {0D5D0CA3-C6B8-436C-82B0-39820C8A13FA}
