--- Tests/Application.Tests/Customers/Queries/GetCustomersListQueryHandlerTests.cs
@@ -1,36 +1,0 @@
-﻿using AutoMapper;
-using Northwind.Application.Customers.Queries.GetCustomersList;
-using Northwind.Application.Tests.Infrastructure;
-using Northwind.Persistence;
-using Shouldly;
-using System.Threading;
-using System.Threading.Tasks;
-using Xunit;
-
-namespace NorthwindTraders.Application.UnitTests.Infrastructure
-{
-    [Collection("QueryCollection")]
-    public class GetCustomersListQueryHandlerTests
-    {
-        private readonly NorthwindDbContext _context;
-        private readonly IMapper _mapper;
-
-        public GetCustomersListQueryHandlerTests(QueryTestFixture fixture)
-        {
-            _context = fixture.Context;
-            _mapper = fixture.Mapper;
-        }
-
-        [Fact]
-        public async Task GetCustomersTest()
-        {
-            var sut = new GetCustomersListQueryHandler(_context, _mapper);
-
-            var result = await sut.Handle(new GetCustomersListQuery(), CancellationToken.None);
-
-            result.ShouldBeOfType<CustomersListViewModel>();
-
-            result.Customers.Count.ShouldBe(3);
-        }
-    }
-}
