--- Tests/Application.Tests/Infrastructure/CommandTestBase.cs
@@ -1,20 +1,0 @@
-using System;
-using Northwind.Persistence;
-
-namespace Northwind.Application.Tests.Infrastructure
-{
-    public class CommandTestBase : IDisposable
-    {
-        protected readonly NorthwindDbContext _context;
-
-        public CommandTestBase()
-        {
-            _context = NorthwindContextFactory.Create();
-        }
-
-        public void Dispose()
-        {
-            NorthwindContextFactory.Destroy(_context);
-        }
-    }
-}
