--- Tests/Application.Tests/MappingTestsFixture.cs
@@ -1,22 +1,0 @@
-﻿using AutoMapper;
-using Northwind.Application.Infrastructure.Mappings;
-
-namespace Northwind.Application.Tests
-{
-    public class MappingTestsFixture
-    {
-        public MappingTestsFixture()
-        {
-            ConfigurationProvider = new MapperConfiguration(cfg =>
-            {
-                cfg.AddProfile<MappingProfile>();
-            });
-
-            Mapper = ConfigurationProvider.CreateMapper();
-        }
-
-        public IConfigurationProvider ConfigurationProvider { get; }
-
-        public IMapper Mapper { get; }
-    }
-}
