+++ Tests/Application.UnitTests/Customers/Commands/CreateCustomer/CreateCustomerCommandTests.cs
@@ -1,0 +1,28 @@
+﻿using MediatR;
+using Moq;
+using Northwind.Application.Customers.Commands.CreateCustomer;
+using System.Threading;
+using Northwind.Application.UnitTests.Infrastructure;
+using Xunit;
+
+namespace Northwind.Application.UnitTests.Customers.Commands.CreateCustomer
+{
+    public class CreateCustomerCommandTests : CommandTestBase
+    {
+        [Fact]
+        public void Handle_GivenValidRequest_ShouldRaiseCustomerCreatedNotification()
+        {
+            // Arrange
+            var mediatorMock = new Mock<IMediator>();
+            var sut = new CreateCustomerCommand.Handler(_context, mediatorMock.Object);
+            var newCustomerId = "QAZQ1";
+
+            // Act
+            var result = sut.Handle(new CreateCustomerCommand { Id = newCustomerId }, CancellationToken.None);
+
+
+            // Assert
+            mediatorMock.Verify(m => m.Publish(It.Is<CustomerCreated>(cc => cc.CustomerId == newCustomerId), It.IsAny<CancellationToken>()), Times.Once);
+        }
+    }
+}
