+++ Tests/Application.UnitTests/Infrastructure/CommandTestBase.cs
@@ -1,0 +1,20 @@
+using System;
+using Northwind.Persistence;
+
+namespace Northwind.Application.UnitTests.Infrastructure
+{
+    public class CommandTestBase : IDisposable
+    {
+        protected readonly NorthwindDbContext _context;
+
+        public CommandTestBase()
+        {
+            _context = NorthwindContextFactory.Create();
+        }
+
+        public void Dispose()
+        {
+            NorthwindContextFactory.Destroy(_context);
+        }
+    }
+}
