+++ Tests/Application.UnitTests/MappingTestsFixture.cs
@@ -1,0 +1,22 @@
+﻿using AutoMapper;
+using Northwind.Application.Infrastructure.Mappings;
+
+namespace Northwind.Application.UnitTests
+{
+    public class MappingTestsFixture
+    {
+        public MappingTestsFixture()
+        {
+            ConfigurationProvider = new MapperConfiguration(cfg =>
+            {
+                cfg.AddProfile<MappingProfile>();
+            });
+
+            Mapper = ConfigurationProvider.CreateMapper();
+        }
+
+        public IConfigurationProvider ConfigurationProvider { get; }
+
+        public IMapper Mapper { get; }
+    }
+}
