+++ Tests/WebUI.IntegrationTests/Common/CustomWebApplicationFactory.cs
@@ -1,0 +1,114 @@
+﻿using System;
+using System.Net.Http;
+using System.Threading.Tasks;
+using IdentityModel.Client;
+using Microsoft.AspNetCore.Hosting;
+using Microsoft.AspNetCore.Mvc.Testing;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Logging;
+using Northwind.Application.Interfaces;
+using Northwind.Persistence;
+
+namespace Northwind.WebUI.IntegrationTests.Common
+{
+    public class CustomWebApplicationFactory<TStartup> : WebApplicationFactory<TStartup> where TStartup : class
+    {
+        protected override void ConfigureWebHost(IWebHostBuilder builder)
+        {
+            builder
+                .ConfigureServices(services =>
+                {
+                    // Create a new service provider.
+                    var serviceProvider = new ServiceCollection()
+                        .AddEntityFrameworkInMemoryDatabase()
+                        .BuildServiceProvider();
+
+                    // Add a database context using an in-memory 
+                    // database for testing.
+                    services.AddDbContext<INorthwindDbContext, NorthwindDbContext>(options =>
+                    {
+                        options.UseInMemoryDatabase("InMemoryDbForTesting");
+                        options.UseInternalServiceProvider(serviceProvider);
+                    });
+
+                    var sp = services.BuildServiceProvider();
+
+                    // Create a scope to obtain a reference to the database
+                    // context (NorthwindDbContext)
+                    using (var scope = sp.CreateScope())
+                    {
+                        var scopedServices = scope.ServiceProvider;
+                        var context = scopedServices.GetRequiredService<INorthwindDbContext>();
+                        var logger = scopedServices.GetRequiredService<ILogger<CustomWebApplicationFactory<TStartup>>>();
+
+                        var concreteContext = (NorthwindDbContext)context;
+
+                        // Ensure the database is created.
+                        concreteContext.Database.EnsureCreated();
+
+                        try
+                        {
+                            // Seed the database with test data.
+                            Utilities.InitializeDbForTests(concreteContext);
+                        }
+                        catch (Exception ex)
+                        {
+                            logger.LogError(ex, $"An error occurred seeding the " +
+                                                "database with test messages. Error: {ex.Message}");
+                        }
+                    }
+                })
+                .UseEnvironment("Test");
+        }
+
+        public HttpClient GetAnonymousClient()
+        {
+            return CreateClient();
+        }
+
+        public async Task<HttpClient> GetAuthenticatedClientAsync()
+        {
+            return await GetAuthenticatedClientAsync("jason@northwind", "Northwind1!");
+        }
+
+        public async Task<HttpClient> GetAuthenticatedClientAsync(string userName, string password)
+        {
+            var client = CreateClient();
+
+            var token = await GetAccessTokenAsync(client, userName, password);
+
+            client.SetBearerToken(token);
+
+            return client;
+        }
+
+        private async Task<string> GetAccessTokenAsync(HttpClient client, string userName, string password)
+        {
+            var disco = await client.GetDiscoveryDocumentAsync();
+
+            if (disco.IsError)
+            {
+                throw new Exception(disco.Error);
+            }
+
+            var response = await client.RequestPasswordTokenAsync(new PasswordTokenRequest
+            {
+                Address = disco.TokenEndpoint,
+                ClientId = "Northwind.IntegrationTests",
+                ClientSecret = "secret",
+
+                Scope = "Northwind.WebUIAPI openid profile",
+                UserName = userName,
+                Password = password
+            });
+
+            if (response.IsError)
+            {
+                throw new Exception(response.Error);
+            }
+
+            return response.AccessToken;
+        }
+    }
+}
