+++ Tests/WebUI.IntegrationTests/Controllers/Customers/GetAll.cs
@@ -1,0 +1,32 @@
+using System.Threading.Tasks;
+using Northwind.Application.Customers.Queries.GetCustomersList;
+using Northwind.WebUI.IntegrationTests.Common;
+using Xunit;
+
+namespace Northwind.WebUI.IntegrationTests.Controllers.Customers
+{
+    public class GetAll : IClassFixture<CustomWebApplicationFactory<Startup>>
+    {
+        private readonly CustomWebApplicationFactory<Startup> _factory;
+
+        public GetAll(CustomWebApplicationFactory<Startup> factory)
+        {
+            _factory = factory;
+        }
+
+        [Fact]
+        public async Task ReturnsCustomersListViewModel()
+        {
+            var client = await _factory.GetAuthenticatedClientAsync();
+
+            var response = await client.GetAsync("/api/customers/getall");
+
+            response.EnsureSuccessStatusCode();
+
+            var vm = await Utilities.GetResponseContent<CustomersListViewModel>(response);
+
+            Assert.IsType<CustomersListViewModel>(vm);
+            Assert.NotEmpty(vm.Customers);
+        }
+    }
+}
